Card items are splited for better maintainablility
